+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntitySearchModel.java
@@ -35,6 +35,7 @@ import is.codion.framework.domain.entity.condition.Condition;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -59,7 +60,9 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
 
 	private final EntityType entityType;
 	private final Collection<Column<String>> columns;
-	private final ValueSet<Entity> entities = ValueSet.valueSet(Notify.WHEN_SET);
+	private final ValueSet<Entity> entities = ValueSet.builder(Collections.<Entity>emptySet())
+					.notify(Notify.WHEN_SET)
+					.build();
 	private final EntityConnectionProvider connectionProvider;
 	private final Map<Column<String>, Settings> settings;
 	private final Value<String> searchString = Value.nonNull("")