+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - EventObserver.addDataListener() renamed addConsumer(), related renaming.
 - Value.Builder added, replaces a couple of static factory methods.
 - Value.Builder.link() added.
+- ValueSet.Builder added, replaces a couple of static factory methods.
 ### is.codion.framework.domain
 - QueryKeyGenerator, SequenceKeyGenerator, now overwrite a current value instead of silently skipping it.
 - AbstractQueriedKeyGenerator.selectAndPopulate() primary key column definition parameter removed.